Move `environment::identify` into `syntactic_environment`
