+++ VERSION
@@ -1,1 +1,1 @@
-0.4.526
+0.4.527
