+++ include/meevax/kernel/environment.hpp
@@ -28,11 +28,11 @@ namespace meevax
 {
 inline namespace kernel
 {
-  class environment : public syntactic_environment
-                    , public configurator<environment>
+  class environment : public configurator<environment>
                     , public machine<environment>
                     , public optimizer
                     , public reader<environment>
+                    , public syntactic_environment<environment>
   {
     using syntactic_environment::syntactic_environment;
 
@@ -64,77 +64,6 @@ inline namespace kernel
 
     auto load(std::string const&) -> object;
 
-    auto identify(object const& variable, object const& scope) const -> object
-    {
-      if (not variable.is_also<identifier>())
-      {
-        return f;
-      }
-      else
-      {
-        for (auto outer = std::begin(scope); outer != std::end(scope); ++outer)
-        {
-          for (auto inner = std::begin(*outer); inner != std::end(*outer); ++inner)
-          {
-            if (inner.get().is<pair>() and (*inner).is<absolute>() and eq((*inner).as<absolute>().symbol(), variable))
-            {
-              return *inner;
-            }
-            else if (inner.get().is<pair>() and eq(*inner, variable))
-            {
-              return make<relative>(make(static_cast<identity::index>(std::distance(std::begin(scope), outer))),
-                                    make(static_cast<identity::index>(std::distance(std::begin(*outer), inner))));
-            }
-            else if (inner.get().is<symbol>() and eq(inner, variable))
-            {
-              return make<variadic>(make(static_cast<identity::index>(std::distance(std::begin(scope), outer))),
-                                    make(static_cast<identity::index>(std::distance(std::begin(*outer), inner))));
-            }
-          }
-        }
-
-        if (variable.is<syntactic_closure>())
-        {
-          return variable.as<syntactic_closure>().identify();
-        }
-        else
-        {
-          return assq(variable, global());
-        }
-      }
-    }
-
-    auto identify(object const& variable, object const& scope)
-    {
-      if (not variable.is_also<identifier>())
-      {
-        return f;
-      }
-      else if (let const& identity = std::as_const(*this).identify(variable, scope); is_truthy(identity))
-      {
-        return identity;
-      }
-      else /* --------------------------------------------------------------------
-      *
-      *  At the outermost level of a program, a definition
-      *
-      *      (define <variable> <expression>)
-      *
-      *  has essentially the same effect as the assignment expression
-      *
-      *      (set! <variable> <expression>)
-      *
-      *  if <variable> is bound to a non-syntax value. However, if <variable> is
-      *  not bound, or is a syntactic keyword, then the definition will bind
-      *  <variable> to a new location before performing the assignment, whereas
-      *  it would be an error to perform a set! on an unbound variable.
-      *
-      * ----------------------------------------------------------------------- */
-      {
-        return car(global() = make<absolute>(variable, undefined) | global());
-      }
-    }
-
     auto operator [](object const&) -> object const&;
 
     auto operator [](std::string const&) -> object const&;
@@ -147,6 +76,8 @@ inline namespace kernel
   extern template class machine<environment>;
 
   extern template class reader<environment>;
+
+  extern template struct syntactic_environment<environment>;
 } // namespace kernel
 } // namespace meevax
